requirements.txt
absl-py==1.3.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==23.3.0
bleach==5.0.1
bokeh==2.4.3
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.4
colorama==0.4.5
contourpy==1.0.6
cvzone==1.5.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.12.4
flatbuffers==22.11.23
fonttools==4.38.0
idna==3.4
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonschema==4.14.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mediapipe==0.9.0
mistune==2.0.4
MouseInfo==0.1.3
mpmath==1.2.1
mypy-extensions==1.0.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
numba==0.56.4
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
packaging==23.1
pandocfilters==1.5.0
panel==0.13.1
param==1.12.2
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.8.1
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
pyct==0.4.8
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pyviz-comms==2.2.1
pywin32==304
pywinpty==2.0.7
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.3.1
QtPy==2.2.0
requests==2.28.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.3.2.post1
stack-data==0.4.0
sympy==1.11.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
txt2tags==3.8
typing_extensions==4.3.0
uri-template==1.3.0
urllib3==1.26.12
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.2
